requirements.txt
# Auto Dev Loop - Dependencies

# Core
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.9.0

# Browser Automation
playwright>=1.40.0

# Dashboard
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0

# Computer Control (for Guardian)
pyautogui>=0.9.54
Pillow>=10.0.0

# Optional: For MCP Support
# mcp>=0.9.0
# anthropic-mcp>=0.1.0

# Optional: For Local Models
# ollama>=0.1.0
